Revert "Remove 4267 suppression from mono (#66552)"
This reverts commit c3dbdea91835d67cc461b22125e652ad5063d746.

This caused `mono-aot-cross.exe` to crash on windows, breaking all the
windows/AOT tests.
Issue: https://github.com/dotnet/runtime/issues/66718

+++ src/mono/mono/metadata/appdomain.c
@@ -634,7 +634,7 @@ real_load (gchar **search_path, const gchar *culture, const gchar *name, const M
 	gchar **path;
 	gchar *filename;
 	const gchar *local_culture;
-	size_t len;
+	gint len;
 
 	if (!culture || *culture == '\0') {
 		local_culture = "";
